Fix bug in homo lr
